添加获取收藏，关注的api

+++ src/main/java/com/guo/technologyforum/dao/entity/vo/ThemeClassNodeVO.java
@@ -11,5 +11,5 @@ import lombok.Data;
 @Data
 public class ThemeClassNodeVO {
     private ThemeClass themeClass;
-    private Integer themeCount;
+    private Long themeCount;
 }
